% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{single_ridge_regression}
\alias{single_ridge_regression}
\title{Runs a ridge regression for a single lambda}
\usage{
single_ridge_regression(dat, response, lambda)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{lambda}{A numeric value of a penalty terms to try}
}
\value{
A data frame of coefficients
}
\description{
Runs a ridge regression for a single lambda
}
